Fix integer overflow in torrent size computation
The computation that validated that the number of pieces announced in
the torrent matched the announced torrent total size was int-based and
could thus overflow for large (> 4GB) files. Fixed the computation to be
long-based to fix this issue.

+++ src/com/turn/ttorrent/client/SharedTorrent.java
@@ -167,7 +167,7 @@ public class SharedTorrent extends Torrent implements PeerActivityListener {
 					.getBytes());
 
 			if (this.piecesHashes.capacity() / Torrent.PIECE_HASH_SIZE *
-					this.pieceLength < this.getSize()) {
+					(long)this.pieceLength < this.getSize()) {
 				throw new IllegalArgumentException("Torrent size does not " +
 						"match the number of pieces and the piece size!");
 			}
